# openapi/valk-dashboard.yaml
openapi: 3.1.0
info:
  title: VALK Dashboard API
  version: 2.0.0
  description: Contracts consumed by the Next.js BFF. Legacy API-key authentication remains supported beside short-lived dashboard bearer tokens.
servers:
  - url: /api
security:
  - dashboardBearer: []
  - legacyApiKey: []
paths:
...
  /dashboard/bgs/rules:
    get:
      operationId: listBgsRules
      security: [{ dashboardBearer: [] }]
      x-capability: rules:write
      responses:
        "200":
          description: Visible personal and tenant BGS rules
          content:
            application/json:
              schema:
                type: object
                required: [data, generated_at]
                properties:
                  data:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/BgsRule" },
                    }
                  generated_at: { type: string, format: date-time }
        default: { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createBgsRule
      security: [{ dashboardBearer: [] }]
      x-capability: rules:write
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/BgsRuleInput" } }
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties: { data: { $ref: "#/components/schemas/BgsRule" } }
        default: { $ref: "#/components/responses/ErrorResponse" }
...
components:
  securitySchemes:
    dashboardBearer: { type: http, scheme: bearer, bearerFormat: JWT }
    legacyApiKey: { type: apiKey, in: header, name: apikey }
...
  schemas:
...
    BgsRuleInput:
      type: object
      required:
        [
          name,
          owner_scope,
          target_scope,
          condition_type,
          threshold_pp,
          window_days,
          severity,
          personal_discord,
          tenant_discord,
          enabled,
        ]
      properties:
        name: { type: string, minLength: 3, maxLength: 160 }
        owner_scope: { type: string, enum: [personal, tenant] }
        target_scope: { type: string, enum: [system, watchlist_all] }
        target_system: { type: [string, "null"], maxLength: 255 }
        condition_type: { $ref: "#/components/schemas/BgsRuleConditionType" }
        condition: { $ref: "#/components/schemas/BgsCondition" }
        threshold_pp: { type: number, exclusiveMinimum: 0, maximum: 100 }
        window_days: { type: integer, minimum: 1, maximum: 30 }
        severity: { type: string, enum: [info, warning, critical] }
        personal_discord: { type: boolean }
        tenant_discord: { type: boolean }
        enabled: { type: boolean }
    BgsRule:
      allOf:
        - { $ref: "#/components/schemas/BgsRuleInput" }
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            owner_user_id: { type: [string, "null"] }
            package_id: { type: [string, "null"], format: uuid }
            template_id: { type: [string, "null"] }
            template_version: { type: [integer, "null"] }
            template_item_key: { type: [string, "null"] }
            effective_from: { type: [string, "null"], format: date-time }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    BgsRuleConditionType:
      type: string
      enum:
        [
          controller_below,
          controller_gap,
          competitor_gain,
          competitor_loss,
          controller_loss,
          tenant_faction_loss,
          tenant_faction_new_conflict,
          tenant_faction_below,
          tenant_faction_gap,
        ]
    BgsCondition:
      type: object
      required: [type]
      properties:
        type: { $ref: "#/components/schemas/BgsRuleConditionType" }
        threshold_pp: { type: number, exclusiveMinimum: 0, maximum: 100 }
        window_days: { type: integer, minimum: 1, maximum: 30 }
        comparison: { type: string, enum: [previous_settled_tick] }
        gap_mode: { type: string, enum: [absolute] }
        conflict_types:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, enum: [election, war] }
...
    ErrorEnvelope:
      {
        type: object,
        required: [error],
        properties:
          {
            error:
              {
                type: object,
                required: [code, message, correlation_id],
                properties:
                  {
                    code: { type: string },
                    message: { type: string },
                    correlation_id: { type: string },
                  },
              },
            generated_at: { type: string, format: date-time },
          },
      }
  responses:
...
    ErrorResponse:
      {
        description: Uniform error envelope,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
          },
      }